refactor types for runners module
